code/requirements.txt
anthropic>=0.30.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
python-dotenv>=1.0.0
rank-bm25>=0.2.2
numpy>=1.24.0
